refactor: extract useOpenClawAttachments and useOpenClawStream hooks from OpenClaw.jsx
Splits the ~1000-line OpenClaw.jsx into three focused files:

- useOpenClawAttachments.js: all attachment state, validation constants,
  drag/drop/paste/select handlers, and object-URL lifecycle management.
- useOpenClawStream.js: message state, SSE streaming, scroll-to-bottom
  RAF effect, abort controller, and loadMessages/handleSend/handleStop.
- OpenClaw.jsx: reduced to page-level state (sessions, status, apps,
  composer, context) and the render tree; sending/messagesError are
  hoisted to the component to share across both hooks without a
  circular dependency.
